ConsoleService/requirements.txt
pydantic==1.5.1
uvicorn==0.11.5
fastapi==0.62.0
tortoise-orm==0.16.18
aiomysql==0.0.20
passlib==1.7.2
orjson==3.1.0
bcrypt==3.1.7
PyJWT==1.7.1
requests==2.23.0
aioredis==1.3.1
rsa
loguru
